skip import when persisted data exists
When running in a kubernetes cluster, the import step should usually only happen once.
For example, this could be done using an init container.
Doing so and changing the deployment in a way that triggers a re-deploy currently fails
because the database has already been initialized and persisted.

To avoid this, write an empty file to the postgresql directory and check whether it exists
on the next successful run to avoid running the setup again.