Add "--break-system-packages" to Docker "pip" command
I'm building a 23.10 based version and they've made changes
(either in Ubuntu or in pip, I'm assuming Ubuntu) so that the
current pip3 command in the Docker file will fail.  For 24.10,
I believe we will need "pip3 install --break-system-packages",
and that appears to be safe to use on 22.04 as well.